refactor: consolidate identity CID tests, filter direct pins
- inline identity CID check (c.Prefix().MhType == mh.IDENTITY) in all
  emit paths: WalkDAG, WalkEntityRoots, and direct pin loops in both
  NewUniquePinnedProvider and NewPinnedEntityRootsProvider
- move all identity CID tests to dag/walker/identity_test.go
- add provider-level identity tests for direct pins and recursive DAGs
